Build and docs updates
* Improve naming consistency in project configuration files, MultiVersio
  now used everywhere.

* Project dependencies are fetched as version/ref pinned git submodules.
  Emilie's stmlib and DaisySP both pinned to current main, while
  libDaisy is pinned to the ref just prior to an AudioCallback signature
  breaking change, pre-1.0.

* Makefile builds deps when necessary.

* gitignore updated, build dir removed from scc.

* linting enabled via `make lint` using Daisy project's clang rules and
  lint accelerator python script. Code not yet passing lint.

* README updated with project and features descriptions. Text is based
  on modwiggler forum post.
