+++ .gitignore
@@ -1,0 +1,19 @@
+
+.DS_Store
+build/
+
+# Editors
+## Visual Studio Code
+.vscode/
+
+## Intellij
+.idea/
+
+## Vim
+# Vim
+[._]*.s[a-w][a-z]
+[._]s[a-w][a-z]
+*.un~
+Session.vim
+.netrwhist
+*~
