--- vs/MultiDelay.vcxproj.filters
@@ -1,43 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <Filter Include="Source files">
-      <UniqueIdentifier>{25bf0424-9e0c-4ebf-9a42-e550795ef895}</UniqueIdentifier>
-      <Extensions>cpp;c;cc;cxx;def;odl;idl;hpj;bat;asm;asmx</Extensions>
-    </Filter>
-    <Filter Include="Header files">
-      <UniqueIdentifier>{e78db3a8-de5c-481e-9304-c32416791e50}</UniqueIdentifier>
-      <Extensions>h;hpp;hxx;hm;inl;inc;xsd</Extensions>
-    </Filter>
-    <Filter Include="Resource files">
-      <UniqueIdentifier>{c6e365c4-e95b-4603-b436-18c976f92be3}</UniqueIdentifier>
-      <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav</Extensions>
-    </Filter>
-    <Filter Include="Source files\Device-specific files">
-      <UniqueIdentifier>{f385e9ba-f4a4-4de5-9139-9c485659e5ca}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="VisualGDB settings">
-      <UniqueIdentifier>{59929191-3401-4c01-828e-d46b49bd7225}</UniqueIdentifier>
-    </Filter>
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="..\MultiEffect.cpp">
-      <Filter>Source files</Filter>
-    </ClCompile>
-    <None Include="stm32.props">
-      <Filter>Source files\Device-specific files</Filter>
-    </None>
-    <None Include="MultiEffect-Debug.vgdbsettings">
-      <Filter>VisualGDB settings</Filter>
-    </None>
-    <None Include="MultiEffect-Release.vgdbsettings">
-      <Filter>VisualGDB settings</Filter>
-    </None>
-    <None Include="stm32.xml">
-      <Filter>VisualGDB settings</Filter>
-    </None>
-    <ClCompile Include="..\..\..\libdaisy\core\startup_stm32h750xx.c">
-      <Filter>Source files\Device-specific files</Filter>
-    </ClCompile>
-  </ItemGroup>
-</Project>
